Delete the regex endpoint util
I'm pretty sure this doesn't work because the regex match result will be shared across different requests :O
I want to YAGNI this until I know what I really need. This sort of thing: doing some regex parsing and use the result in both the when and do, can be achieved by making it middleware, but it might be bad if you have loads of middleware for regexing at every endpoint. Might want middleware defined on the level of an endpoint, but having loads of subquelaags might be annoying.
Could make addEndpoint optionally take a function that produces an endpoint so you can have the kind of class I've deleted actually work by having the data per request handle.